tests/wpt/tests/navigation-api/updateCurrentEntry-method/cross-document-location-api.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    i.contentWindow.navigation.updateCurrentEntry({ state: { data: "value" } });
    assert_equals(i.contentWindow.navigation.entries().length, 1);
    assert_equals(i.contentWindow.navigation.currentEntry.getState().data, "value");

    i.contentWindow.location.href = "?1";
    i.onload = t.step_func_done(() => {
      assert_equals(i.contentWindow.navigation.entries().length, 2);
      assert_equals(i.contentWindow.navigation.currentEntry.index, 1);
      assert_equals(i.contentWindow.navigation.currentEntry.getState(), undefined);
    });
  });
}, "entry.getState() behavior after cross-document location API navigation");
</script>
