LayoutTests/imported/w3c/web-platform-tests/navigation-api/updateCurrentEntry-method/cross-document-location-api-expected.txt



FAIL entry.getState() behavior after cross-document location API navigation assert_equals: expected 2 but got 1
